+++ src/main/java/com/boris/fundingarbitrage/exchange/impl/gate/publicrest/GatePublicHttpClient.java
@@ -8,6 +8,7 @@ import com.boris.fundingarbitrage.model.contract.FundingRate;
 import com.boris.fundingarbitrage.util.https.PrettyHttpClient;
 import com.boris.fundingarbitrage.util.https.RequestProcessingClientWrapper;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
@@ -30,17 +31,15 @@ public class GatePublicHttpClient extends PublicHttpClient {
 
 	@Override
 	protected CompletableFuture<Map<String, PublicOnePullData>> getPublicOnePullData() {
-		CompletableFuture<PublicResponses.ContractsResponse> contractsResponseFuture = requestWrapper.getResponse(
-						PublicEndpoints.contractsRequestSymbols(),
+		CompletableFuture<PublicResponses.ContractsResponse> contractsResponseFuture = requestWrapper.getResponse(PublicEndpoints.contractsRequestSymbols(),
 						PublicResponses.ContractsResponse.class
 		);
-		CompletableFuture<PublicResponses.TickersResponse> tickersResponseFuture = requestWrapper.getResponse(
-						PublicEndpoints.tickersRequestSymbols(),
+		CompletableFuture<PublicResponses.TickersResponse> tickersResponseFuture = requestWrapper.getResponse(PublicEndpoints.tickersRequestSymbols(),
 						PublicResponses.TickersResponse.class
 		);
 
 		return CompletableFuture.allOf(contractsResponseFuture, tickersResponseFuture).thenApply(_ -> {
-			Map<String, Double> lotSizes = contractsResponseFuture.join().getLotSizes();
+			Map<String, BigDecimal> lotSizes = contractsResponseFuture.join().getLotSizes();
 			Map<String, Integer> fundingGranularityHours = contractsResponseFuture.join().getFundingGranularityHours();
 			Map<String, BookTicker> bookTickers = tickersResponseFuture.join().getBookTickers();
 			Map<String, Double> volumes24h = tickersResponseFuture.join().getVolume24h();
